+++ main.cpp
@@ -1,18 +1,18 @@
+// SPDX-FileCopyrightText: 2003-2014 Jesper K. Pedersen <blackie@kde.org>
 // SPDX-FileCopyrightText: 2003 Lukáš Tinkl <lukas@kde.org>
 // SPDX-FileCopyrightText: 2003 Simon Hausmann <hausmann@kde.org>
 // SPDX-FileCopyrightText: 2003 Stephan Binner <binner@kde.org>
-// SPDX-FileCopyrightText: 2003-2014 Jesper K. Pedersen <blackie@kde.org>
 // SPDX-FileCopyrightText: 2005-2007 Dirk Mueller <mueller@kde.org>
 // SPDX-FileCopyrightText: 2006-2008 Tuomas Suutari <tuomas@nepnep.net>
-// SPDX-FileCopyrightText: 2007 Chusslove Illich <caslav.ilic@gmx.net>
 // SPDX-FileCopyrightText: 2007-2011 Jan Kundrát <jkt@flaska.net>
+// SPDX-FileCopyrightText: 2007 Chusslove Illich <caslav.ilic@gmx.net>
 // SPDX-FileCopyrightText: 2009 Andrew Coles <andrew.i.coles@googlemail.com>
 // SPDX-FileCopyrightText: 2009 Christoph Feck <cfeck@kde.org>
 // SPDX-FileCopyrightText: 2010-2012 Miika Turkia <miika.turkia@gmail.com>
-// SPDX-FileCopyrightText: 2012-2024 Johannes Zarl-Zierl <johannes@zarl-zierl.at>
+// SPDX-FileCopyrightText: 2012-2025 Johannes Zarl-Zierl <johannes@zarl-zierl.at>
+// SPDX-FileCopyrightText: 2018-2024 Tobias Leupold <tl@stonemx.de>
 // SPDX-FileCopyrightText: 2018 Antoni Bella Pérez <antonibella5@yahoo.com>
 // SPDX-FileCopyrightText: 2019 Alexander Potashev <aspotashev@gmail.com>
-// SPDX-FileCopyrightText: 2018-2024 Tobias Leupold <tl@stonemx.de>
 //
 // SPDX-License-Identifier: GPL-2.0-or-later
 
@@ -135,6 +135,8 @@ int main(int argc, char **argv)
     if (app.styleSheet().isEmpty())
         app.setStyleSheet(STYLE);
     app.setAttribute(Qt::AA_UseStyleSheetPropagationInWidgetStyles);
+
+    qDebug() << "kphotoalbumrc:" << KConfig::mainConfigName();
 
     new MainWindow::SplashScreen();