+++ testcases/integration-tests.sh
@@ -1,10 +1,10 @@
 #!/bin/bash
 
-# SPDX-FileCopyrightText: 2018-2021 Johannes Zarl-Zierl <johannes@zarl-zierl.at>
+# SPDX-FileCopyrightText: 2018 - 2025 Johannes Zarl-Zierl <johannes@zarl-zierl.at>
 #
 # SPDX-License-Identifier: BSD-2-Clause
 
-export LC_ALL=C
+export LC_ALL=C.UTF-8
 myname=`basename "$0"`
 mydir=`dirname "$0"`