+++ testcases/integration-tests/check_autostacking.sh
@@ -1,4 +1,4 @@
-# SPDX-FileCopyrightText: 2018-2020 Johannes Zarl-Zierl <johannes@zarl-zierl.at>
+# SPDX-FileCopyrightText: 2018 - 2025 Johannes Zarl-Zierl <johannes@zarl-zierl.at>
 #
 # SPDX-License-Identifier: BSD-2-Clause
 
@@ -33,7 +33,7 @@ prepare_check_autostacking()
 call_check_autostacking()
 {
 	local check_dir="$1"
-	kphotoalbum --db "$check_dir/db/index.xml" --search
+	kphotoalbum --config "$check_dir/kphotoalbumrc" --db "$check_dir/db/index.xml" --search
 }
 check_autostacking()
 {